requirements-streamlit.txt
streamlit>=1.24.0
pandas>=1.5
numpy>=1.24
tldextract>=3.4
google-generativeai>=0.3.0
httpx>=0.24.0
requests>=2.28
joblib>=1.2
scikit-learn>=1.0
python-dotenv>=1.0
pytest>=7.0